Split out `HumanVisitor` into its own file.
