+++ src/loggy-intf/package.json
@@ -10,6 +10,7 @@
   },
   "imports": {
     "#x/*": "./export/*.js",
+    "#p/*": "./private/*.js",
     "#tests/*": "./tests/*.js"
   },
